requirements.txt
python-dotenv
pandas
tqdm
requests
pyyaml

# Hosted Llama-70B via Groq
groq

# Self-hosted / local backends
transformers
accelerate
sentencepiece
